[PROTO-04] 던전 팝업 창 및 유닛 탐색 기능 구현
- UI Manager와 UI 관련 컨트롤러간 연결 로직 개선
- 던전 정보 (탐색) 팝업 및 유닛 선택 창 구현
- 단순하게 유닛을 던전에 보낼 수 있도록 구현
- 임시 리소스 데이타 존재

+++ Assets/Days/Game/Object/Infra/ObjectInfo.cs
@@ -8,7 +8,7 @@ namespace Days.Game.Object.Infra
 {
     public class ObjectInfo
     {
-        public ushort Index { get; set; }
+        public byte Index { get; set; }
         public string Name { get; set; }
         public State State { get; set; }
         public Character Character { get; set; }
